pico hand tracking setting
pico requires hand tracking enabled to launch. it's hidden in a menu inside the feature groups. added it to the setup menu. The other settings are more performance-oriented, so I think that's more of a developer choice. But it could be cool to make another menu that guides the developer through all the many performance options across different packages.

Maybe an 'optional' segment